fix(ci): rustfmt and clippy warnings in API crate
Apply cargo fmt to CCTP modules and resolve collapsible_match,
unnecessary cast, and manual_filter clippy lints.

+++ crates/api/src/cctp/stellar_mint_verifier.rs
@@ -343,7 +343,7 @@ impl StellarMintVerifier for StellarRpcMintVerifier {
             message,
             nonce,
             recipient,
-            amount_cctp as u128,
+            amount_cctp,
             quoted_finality,
         )
         .await
